refactor: move FolderTree to resources package
